Create Simple Fun #160: Cut The Ropes
This code is somehow full mess but it still works and passed all tests. The most amusing thing is lens.pop() at the end of the code. Before When I run the code it returned an array with unnecessary 0 at the end, and I came up with thsi so called "creative solution"